refactor: extract Twilio settings related components to own directory
